Link from category to details page

+++ HornSneaker/src/main/java/vn/edu/hcmus/hornsneaker/service/ProductServices.java
@@ -21,7 +21,7 @@ public class ProductServices {
 	}
 	*/
 
-	public ProductEntity find(int id) {
-		return productRepo.getById(Long.valueOf(id));		
+	public ProductEntity find(Long id) {
+		return productRepo.getById(id);		
 	}
 }
